+++ src/database/database.cpp
@@ -119,14 +119,16 @@ db::IrcChannelOptions Database::get_irc_channel_options_with_server_and_global_d
   return coptions;
 }
 
-void Database::store_muc_message(const std::string& owner, const Iid& iid,
-                                 Database::time_point date,
-                                 const std::string& body,
-                                 const std::string& nick)
+std::string Database::store_muc_message(const std::string& owner, const Iid& iid,
+                                        Database::time_point date,
+                                        const std::string& body,
+                                        const std::string& nick)
 {
   db::MucLogLine line(*Database::db);
 
-  line.uuid = Database::gen_uuid();
+  auto uuid = Database::gen_uuid();
+
+  line.uuid = uuid;
   line.owner = owner;
   line.ircChanName = iid.get_local();
   line.ircServerName = iid.get_server();
@@ -135,6 +137,8 @@ void Database::store_muc_message(const std::string& owner, const Iid& iid,
   line.nick = nick;
 
   line.update();
+
+  return uuid;
 }
 
 std::vector<db::MucLogLine> Database::get_muc_logs(const std::string& owner, const std::string& chan_name, const std::string& server,