+++ src/xmpp/xmpp_component.hpp
@@ -33,6 +33,7 @@
 #define DATAFORM_NS      "jabber:x:data"
 #define RSM_NS           "http://jabber.org/protocol/rsm"
 #define MUC_TRAFFIC_NS   "http://jabber.org/protocol/muc#traffic"
+#define STABLE_ID_NS     "urn:xmpp:sid:0"
 
 /**
  * An XMPP component, communicating with an XMPP server using the protocole
@@ -134,7 +135,8 @@ public:
   /**
    * Send a (non-private) message to the MUC
    */
-  void send_muc_message(const std::string& muc_name, const std::string& nick, Xmpp::body&& body, const std::string& jid_to);
+  void send_muc_message(const std::string& muc_name, const std::string& nick, Xmpp::body&& body, const std::string& jid_to,
+                        std::string uuid);
   /**
    * Send a message, with a <delay/> element, part of a MUC history
    */