+++ src/utils/string.hpp
@@ -6,5 +6,3 @@
 
 bool to_bool(const std::string& val);
 std::vector<std::string> cut(const std::string& val, const std::size_t size);
-
-