+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Version 5.0
  - Use the udns library instead of c-ares, for asynchronous DNS resolution.
    It’s still fully optional.
  - Update MAM implementation to version 6.0 (namespace mam:2)
+ - If the client doesn’t specify any limit, MAM results contain at most 100
+   messages, instead of the whole archive.
  - Multiline topics are now properly handled
 
 Version 4.1 - 2017-03-21