+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+Version 4.2 - 2017-04-26
+========================
+
+ - Fix a build issue when LiteSQL is absent from the system
+
 Version 4.1 - 2017-03-21
 ========================